+++ src/MeasureChoice.h
@@ -1,3 +1,25 @@
+//*****************************************************************************
+// The JAZZ++ Midi Sequencer
+//
+// Copyright (C) 1994-2000 Andreas Voss and Per Sigmond, all rights reserved.
+// Modifications Copyright (C) 2004 Patrick Earl
+// Modifications Copyright (C) 2008 Peter J. Stieber
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program; if not, write to the Free Software
+// Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+//*****************************************************************************
+
 #ifndef JZ_MEASURECHOICE_H
 #define JZ_MEASURECHOICE_H
 
@@ -5,27 +27,29 @@
 
 class JZSong;
 
+//*****************************************************************************
+//*****************************************************************************
 class tMeasureChoice : public tNamedValueChoice
 {
   public:
 
     tMeasureChoice(wxWindow* pParent);
 
-   long GetMeasure();
+    int GetMeasure();
 
-   void SetMeasure(long Measure);
+    void SetMeasure(int Measure);
 
-   long GetTicks(JZSong* pSong);
+    int GetTicks(JZSong* pSong);
 };
 
 inline
-long tMeasureChoice::GetMeasure()
+int tMeasureChoice::GetMeasure()
 {
   return GetValue();
 }
 
 inline
-void tMeasureChoice::SetMeasure(long Measure)
+void tMeasureChoice::SetMeasure(int Measure)
 {
   SetValue(Measure);
 }
